Move sample-spec-helper module into the test/helpers subdirectory
The new directory name is more descriptive of its purpose.
