+++ test/sample-notification.spec.js
@@ -1,4 +1,4 @@
-var sampleSpecHelper = require('./modules/sample-spec-helper.js');
+var sampleSpecHelper = require('./helpers/sample-spec-helper.js');
 var testHelper = require('../src/testing/test-helper.js');
 var errorFormatter = testHelper.validationErrorFormatter;
